+++ components/payers/financials.tsx
@@ -8,9 +8,9 @@ import { formatDate, formatRevenue } from "@/helpers/functional";
 import { ColumnDef } from "@tanstack/react-table";
 import { flow } from "lodash/fp";
 import { FC, useEffect, useState } from "react";
-import AnalyticsTable from "../analytics/analytics-table";
-import { AccountMrr } from "../analytics/analytics-table-column";
 import MrrFilterBtnGrp from "../analytics/mrr-filter-btn-grp";
+import AnalyticsTable from "../analytics/table/analytics-table";
+import { AccountMrr } from "../analytics/table/analytics-table-column";
 import { useMrrFilter, withMrrFilter } from "../analytics/useMrrFilter";
 import DefaultAccordionItem from "../ui-elements/accordion/DefaultAccordionItem";
